feat(events): add known event types and factory helpers
Add a centralized registry of known event type strings and a set of
typed factory helpers for creating event objects.

- Introduce packages/storage/src/repositories/events/knownEventTypes.ts
  with a const map of all event type identifiers (accounts, users,
  gardens, transactions, invoices, receipts, raisedBeds, raisedBedFields,
  operations, delivery).
- Add packages/storage/src/repositories/events/knownEvents.ts that
  exports typed helper functions (knownEvents) for constructing events
  with proper type, version, aggregateId and payload. Helpers are
  provided for account, user, garden, transaction, invoice, receipt,
  raised bed and delivery related events, among others.
- Wire payload TypeScript types from ./types to ensure event payloads
  are correctly typed.

This centralizes event naming and reduces repeated string literals,
improves type safety when creating events, and makes future event
versioning easier to manage.
